+++ CHANGELOG.md
@@ -19,6 +19,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   `Hvad er betydningen af følgende talemåde: `.
 - Add `fbgemm-gpu` to `generative` dependencies, as it is required to load newer Llama
   models.
+- When a generative model isn't stored as safetensors, we now report an unknown number
+  of parameters, and log a warning to the user on how to fix this.
 
 
 ## [v14.0.4] - 2024-12-17